refactor(docs): split overview route for its loading skeleton
