+++ apps/docs/src/app/(docs)/packages/[packageId]/[versionId]/[...entitySegments]/loading.tsx
